# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EMMA API
  version: 0.1.0
  description: OpenAPI specification for the EMMA API (developer-docs snapshot)
servers:
  - url: http://localhost:5262
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactCreateDto:
      type: object
      required: [organizationId, firstName, lastName]
      properties:
        organizationId: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string, nullable: true }
        preferredName: { type: string, nullable: true }
        title: { type: string, nullable: true }
        jobTitle: { type: string, nullable: true }
        company: { type: string, nullable: true }
        department: { type: string, nullable: true }
        source: { type: string, nullable: true }
        ownerId: { type: string, format: uuid, nullable: true }
        preferredContactMethod: { type: string, nullable: true }
        preferredContactTime: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        profilePictureUrl: { type: string, nullable: true }
    ContactReadDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string, nullable: true }
        preferredName: { type: string, nullable: true }
        title: { type: string, nullable: true }
        jobTitle: { type: string, nullable: true }
        company: { type: string, nullable: true }
        department: { type: string, nullable: true }
        source: { type: string, nullable: true }
        ownerId: { type: string, format: uuid, nullable: true }
        preferredContactMethod: { type: string, nullable: true }
        preferredContactTime: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        profilePictureUrl: { type: string, nullable: true }
        lastContactedAt: { type: string, format: date-time, nullable: true }
        nextFollowUpAt: { type: string, format: date-time, nullable: true }
    ContactAssignDto:
      type: object
      required: [contactId, userId, assignedByAgentId]
      properties:
        contactId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        assignedByAgentId: { type: string, format: uuid }
        sourceContext: { type: string, nullable: true }
        traceId: { type: string, nullable: true }
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
paths:
...
  /api/Contact:
    post:
      summary: Create contact
      tags: [Contact]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreateDto'
            example:
              organizationId: "11111111-1111-1111-1111-111111111111"
              firstName: "Ada"
              lastName: "Lovelace"
              jobTitle: "Engineer"
              ownerId: "22222222-2222-2222-2222-222222222222"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactReadDto'
              example:
                id: "33333333-3333-3333-3333-333333333333"
                organizationId: "11111111-1111-1111-1111-111111111111"
                firstName: "Ada"
                lastName: "Lovelace"
                jobTitle: "Engineer"
                ownerId: "22222222-2222-2222-2222-222222222222"
        '400': { description: Bad Request }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    get:
      summary: List contacts by organization
      tags: [Contact]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: orgId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactReadDto'
              example:
                - id: "33333333-3333-3333-3333-333333333333"
                  organizationId: "11111111-1111-1111-1111-111111111111"
                  firstName: "Ada"
                  lastName: "Lovelace"
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Contact/{id}:
    get:
      summary: Get contact by id
      tags: [Contact]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactReadDto'
        '404': { description: Not Found }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
  /api/Contact/{id}/assign:
    put:
      summary: Assign contact to user
      tags: [Contact]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactAssignDto'
            example:
              contactId: "33333333-3333-3333-3333-333333333333"
              userId: "22222222-2222-2222-2222-222222222222"
              assignedByAgentId: "99999999-9999-9999-9999-999999999999"
              sourceContext: "crm:bulk-assign"
      responses:
        '204': { description: No Content }
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
              example:
                title: "Validation failed"
                status: 400
                detail: "Path id must match body.contactId."
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
...